# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM для Telegram ботов API",
        "description": "API документация для CRM системы управления Telegram ботами",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1"
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация нового пользователя",
                "description": "Создает нового пользователя и возвращает токен доступа",
                "operationId": "1df5a2272d28ea34b0c54ee361691294",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя пользователя",
                                        "type": "string",
                                        "example": "Иван Иванов"
                                    },
                                    "email": {
                                        "description": "Email пользователя",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "description": "Пароль (минимум 8 символов, буквы разного регистра, цифры, символы)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    },
                                    "password_confirmation": {
                                        "description": "Подтверждение пароля",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    },
                                    "role_id": {
                                        "description": "ID роли (опционально, по умолчанию 1 - User)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешная регистрация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "accessToken": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGc..."
                                        },
                                        "notify": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Успешно"
                                                },
                                                "text": {
                                                    "type": "string",
                                                    "example": "Вы успешно зарегистрировались"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "notify": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Ошибка регистрации"
                                                },
                                                "text": {
                                                    "type": "string",
                                                    "example": "Произошла ошибка на сервере. Попробуйте позже."
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Вход в систему",
                "description": "Авторизует пользователя и возвращает токен доступа",
                "operationId": "0946da348acc1471c3b8508f616cbaea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email пользователя",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "description": "Пароль",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная авторизация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "accessToken": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGc..."
                                        },
                                        "notify": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Успешно"
                                                },
                                                "text": {
                                                    "type": "string",
                                                    "example": "Добро пожаловать!"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "notify": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Доступ запрещен"
                                                },
                                                "text": {
                                                    "type": "string",
                                                    "example": "Неверный email или пароль"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Получить данные текущего пользователя",
                "description": "Возвращает информацию об авторизованном пользователе",
                "operationId": "165a488cb7bb6a7a34ea90ad04b995db",
                "responses": {
                    "200": {
                        "description": "Данные пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите токен в формате: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}